develop-20230904 1. 增加了 redis 服务相关的配置 2. 增加了 redis 服务在 favorite 模块中的一些工具代码 3. 给 favorite action、favorite list、获取用户点赞视频数、获取作者获赞数、获取视频点赞量 等添加了缓存的优化 4. 将 feed 的 jwt 鉴权取消，现在可以离线获取视频流

+++ cmd/favorite/service/favorite_action.go
@@ -2,6 +2,7 @@ package service
 
 import (
 	"context"
+	"github.com/Yra-A/Douyin_Simple_Demo/pkg/errno"
 
 	"github.com/Yra-A/Douyin_Simple_Demo/cmd/favorite/dal/db"
 	"github.com/Yra-A/Douyin_Simple_Demo/kitex_gen/favorite"
@@ -19,11 +20,22 @@ func NewFavoriteActionService(ctx context.Context) *FavoriteActionService {
 }
 
 func (s *FavoriteActionService) FavoriteAction(req *favorite.FavoriteActionRequest) error {
-	//1点赞
+	var err error
+	// 查询视频是否存在
+	_, err = db.CheckVideoExistById(req.VideoId)
+	if err != nil {
+		return err
+	}
+	// 参数出错
+	if req.ActionType != 1 && req.ActionType != 2 {
+		return errno.ParamErr
+	}
+
+	// 1 点赞
 	if req.ActionType == 1 {
 		return db.Add(s.ctx, req)
 	} else {
-		//2取消点赞
+		// 2 取消点赞
 		return db.Delete(s.ctx, req)
 	}
